Moved character-selection to it's own component
